+++ source/test/F0.CodeAnalysis.Testing/CodeAnalysis/CodeRefactorings/CodeRefactoringVerifier.cs
@@ -65,6 +65,13 @@ namespace F0.Testing.CodeAnalysis.CodeRefactorings
 			return tester.RunAsync(CancellationToken.None);
 		}
 
+		public Task CodeActionAsync(string initialCode, string expectedCode, string[][] additionalProjects, LanguageVersion languageVersion)
+		{
+			var assertions = CreateAssertions(initialCode, expectedCode, additionalProjects, languageVersion);
+
+			return assertions.RunAsync(CancellationToken.None);
+		}
+
 		private static CodeRefactoringTester<TCodeRefactoring> CreateTester(string initialCode, string? expectedCode = null, LanguageVersion? languageVersion = null)
 		{
 			var normalizedInitialCode = initialCode.Untabify();
@@ -81,6 +88,28 @@ namespace F0.Testing.CodeAnalysis.CodeRefactorings
 			}
 
 			return tester;
+		}
+
+		private static CodeRefactoringAssertions<TCodeRefactoring> CreateAssertions(string initialCode, string expectedCode, string[][] additionalProjects, LanguageVersion? languageVersion = null)
+		{
+			var testCode = initialCode.Untabify();
+			var fixedCode = expectedCode.Untabify();
+			var assertions = new CodeRefactoringAssertions<TCodeRefactoring>(testCode, fixedCode)
+			{
+				LanguageVersion = languageVersion
+			};
+
+			for (var projectIndex = 0; projectIndex < additionalProjects.Length; projectIndex++)
+			{
+				for (var documentIndex = 0; documentIndex < additionalProjects[projectIndex].Length; documentIndex++)
+				{
+					additionalProjects[projectIndex][documentIndex] = additionalProjects[projectIndex][documentIndex].Untabify();
+				}
+			}
+
+			assertions.AdditionalProjects.AddRange(additionalProjects);
+
+			return assertions;
 		}
 	}
 }